+++ src/main/java/com/trxyzng/trung/utility/EmailUtils.java
@@ -1,4 +1,4 @@
-package com.trxyzng.trung.authentication.shared.utility;
+package com.trxyzng.trung.utility;
 
 import javax.mail.*;
 import javax.mail.internet.InternetAddress;
